added additional check on credentials
we often get an error invalid_grant when getting the google calendar items.

This is decribed in detail here:
https://code.google.com/p/google-api-python-client/issues/detail?id=90

we're guiding the code to get the a new access token when the refresh token is
expired (in the assumption that that is indeed what causes the error). To be
seen.